move : equipment는 작업이 끝났으니 파일 정리
작업 예정
보석 정보 가져오기
UI 배치
